wip gas tracker
