+++ vm/default_runtime/src/gas_block_store.rs
@@ -1,0 +1,11 @@
+use ipld_blockstore::BlockStore;
+use vm::{GasTracker, PriceList};
+
+pub(crate) struct GasBlockStore<'bs>
+where
+    BS: BlockStore,
+{
+    pub price_list: PriceList,
+    pub gas: GasTracker,
+    pub store: &'bs BS,
+}
